chore: wrapped the base component with ThemeProvider at the root level to make theme props available globally
